removed SPI, echo response check, added debug led
echo response check is removed given the wired means means the risk of
lost packets is negligible.

SPI is now not disabled since it is easy to brick your device. revert this
change when you're sure it's the final change.

turned on thed ebug led when the device is sleeping.